+++ Runtime/Scripts/GlobalTools/Utilities/ObjectToggle.cs
@@ -3,10 +3,16 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace PolytopeSolutions.Toolset.GlobalTools.Utilities {
-	public class ObjectToggle : MonoBehaviour {
-		public void Toggle() {
-			gameObject.SetActive(!gameObject.activeSelf);
-		}
-	}
+    public class ObjectToggle : MonoBehaviour {
+        [SerializeField] private bool startActive = true;
+
+        private void Start() {
+            gameObject.SetActive(this.startActive);
+        }
+
+        public void Toggle() {
+            gameObject.SetActive(!gameObject.activeSelf);
+        }
+    }
 }