+++ Runtime/Scripts/UI/UIState.cs
@@ -14,11 +14,18 @@ namespace PolytopeSolutions.Toolset.UI {
     [Serializable]
     public class UIState { // ScriptableObject
         [SerializeField] protected string stateName;
+        [SerializeField] protected string stateID;
         [SerializeField] protected List<Canvas> canvases;
 
-        protected bool isActive = false;
+        public string StateName => this.stateName;
+        public string StateID => this.stateID;
 
-        public virtual void Activate(bool immediate = false) {
+        protected bool isActive = false;
+        public bool Contains(Canvas canvas) {
+            return this.canvases.Contains(canvas);
+        }
+
+        public virtual void Activate(UIState previous=null, bool immediate = false) {
             if (!this.isActive) { 
                 #if DEBUG2
                 this.Log($"Activating {this.name}");
@@ -26,12 +33,19 @@ namespace PolytopeSolutions.Toolset.UI {
                 this.isActive = true;
             }
         }
-        public virtual void Deactivate(bool immediate = false) {
+        public virtual void Deactivate(UIState previous=null, bool immediate = false) {
             if (this.isActive) { 
                 #if DEBUG2
                 this.Log($"Deactivating {this.name}");
                 #endif
                 this.isActive = false;
+            }
+        }
+        public virtual void Toggle() {
+            if (this.isActive) {
+                this.Deactivate();
+            } else {
+                this.Activate();
             }
         }
     }